Move token to own file
