+++ apps/web/src/app/admin-console/organizations/policies/index.ts
@@ -1,2 +1,3 @@
 export { PoliciesComponent } from "./policies.component";
-export { policyRegister, POLICY_REGISTER_TOKEN } from "./policy-register";
+export { policyRegister } from "./policy-register";
+export { POLICY_REGISTER_TOKEN } from "./policy-register-token";
